Refactor: Decouple Core Transformer Blocks
